# api/openapi.yaml
openapi: 3.0.3
info:
  title: Karvon Scraper API
  version: 1.0.0
  description: |
    REST API for Karvon: the scraper module (Google Maps lead scraping with website
    email crawling) and the campaign module (outreach campaigns, contacts and their
    lifecycle, reusable email content, consent and the two delivery providers). Live
    job progress is delivered over Server-Sent Events.

    All routes except `/healthz`, `/openapi.json` and `/webhooks/*` require
    `Authorization: Bearer <KARVON_API_KEY>`. The `/webhooks/*` routes carry no API
    key: they are called by Instantly and Mailchimp, and are authenticated instead by
    the per-registration token in their path together with the secret that provider
    signs its deliveries with.
servers:
  - url: /api/v1
    description: Default mount point
security:
  - bearerAuth: []
...
paths:
...
  /jobs/{id}/export.csv:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [jobs]
      operationId: exportJobCsv
      summary: Stream this job's businesses and emails as CSV
      responses:
        '200':
          description: CSV attachment
          content:
            text/csv:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /businesses:
    get:
      tags: [businesses]
      operationId: listBusinesses
      summary: Master business list
      parameters:
        - name: job_id
          in: query
          schema: { type: string, format: uuid }
        - name: category
          in: query
          schema: { type: string, maxLength: 200 }
        - name: state
          in: query
          schema: { type: string, maxLength: 100 }
        - name: city
          in: query
          schema: { type: string, maxLength: 200 }
        - name: has_email
          in: query
          schema: { type: boolean }
        - name: suppressed
          in: query
          description: Defaults to false; pass true for only suppressed rows
          schema: { type: boolean }
        - name: q
          in: query
          description: Trigram search over name and domain
          schema: { type: string, maxLength: 200 }
        - name: email_source
          in: query
          schema: { $ref: '#/components/schemas/EmailSource' }
        - name: verification_tag
          in: query
          description: Repeatable filter on the tag of the business's primary address
          schema:
            type: array
            items: { $ref: '#/components/schemas/VerificationTag' }
          style: form
          explode: true
        - name: sort
          in: query
          schema:
            type: string
            default: created_at:desc
            enum: [created_at:desc, created_at:asc, name:asc, name:desc, city:asc, city:desc, state:asc, state:desc, rating:desc, rating:asc, reviews:desc, reviews:asc, verification_score:desc, verification_score:asc]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated businesses
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /businesses/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [businesses]
      operationId: getBusiness
      summary: Business detail with every known email
      responses:
        '200':
          description: Business
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [businesses]
      operationId: updateBusiness
      summary: Update suppression flag or notes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BusinessUpdate' }
      responses:
        '200':
          description: Updated business
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /businesses/bulk:
    post:
      tags: [businesses]
      operationId: bulkUpdateBusinesses
      summary: Suppress or unsuppress many businesses at once
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BusinessBulkRequest' }
      responses:
        '200':
          description: Number of rows changed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /businesses/export:
    post:
      tags: [businesses]
      operationId: exportBusinesses
      summary: Stream a filtered business set as CSV
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BusinessExportRequest' }
      responses:
        '200':
          description: CSV attachment
          content:
            text/csv:
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /campaign-analytics/overview:
    get:
      tags: [campaign-analytics]
      operationId: getCampaignOverview
      summary: Campaign counts, the funnel and the overall metrics for the whole area
      responses:
        '200':
          description: Overview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CampaignOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /campaign-analytics/campaigns/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [campaign-analytics]
      operationId: getCampaignAnalytics
      summary: One campaign's metrics beside Instantly's, with the disagreements called out
      description: |
        Our numbers and Instantly's are reported side by side and never merged: the two
        count differently, and `mismatch` treats the gap as a signal rather than
        hiding it behind an average.
      parameters:
        - name: days
          in: query
          description: Length of the daily series
          schema: { type: integer, minimum: 1, maximum: 365, default: 30 }
      responses:
        '200':
          description: Campaign analytics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CampaignAnalytics' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /campaign-analytics/campaigns/{id}/variants:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [campaign-analytics]
      operationId: getCampaignVariantAnalytics
      summary: How each variant of this campaign performed
      responses:
        '200':
          description: Per-variant metrics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/VariantAnalytics' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /campaign-analytics/components:
    get:
      tags: [campaign-analytics]
      operationId: getComponentAnalytics
      summary: How each component performed across the sends that carried it
      parameters:
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/ComponentType' }
        - name: campaign_id
          in: query
          schema: { type: string, format: uuid }
        - name: since
          in: query
          description: Only sends at or after this instant
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Per-component metrics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ComponentAnalytics' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /campaign-analytics/sending-accounts:
    get:
      tags: [campaign-analytics]
      operationId: getAccountAnalytics
      summary: Per-mailbox deliverability, ours beside Instantly's
      responses:
        '200':
          description: Per-account metrics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AccountAnalytics' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /campaign-analytics/funnel:
    get:
      tags: [campaign-analytics]
      operationId: getFunnelAnalytics
      summary: The cold → subscriber funnel, optionally for one campaign
      description: |
        `current` is how many contacts sit in the stage now; `ever_reached` how many
        have passed through it. The two differ because a contact only moves forward,
        so the second is the one to read as a conversion.
      parameters:
        - name: campaign_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Funnel steps
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FunnelStep' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Static API key, `Authorization: Bearer <KARVON_API_KEY>`"
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: Request body or query failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string }
            details:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
      additionalProperties: false
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    PageMeta:
      type: object
      required: [page, per_page, total]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer, format: int64 }
...
    EmailSource:
      type: string
      enum: [mailto, regex, provider]
    BusinessEmail:
      type: object
      required: [id, email, source, is_primary, found_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        source: { $ref: '#/components/schemas/EmailSource' }
        page_url: { type: string, nullable: true }
        is_primary: { type: boolean }
        verification_id: { type: string, format: uuid, nullable: true }
        verified_status:
          allOf:
            - $ref: '#/components/schemas/VerificationTag'
          nullable: true
        verification_score: { type: integer, nullable: true }
        typo_suggestion: { type: string, nullable: true }
        found_at: { type: string, format: date-time }
    Business:
      type: object
      required: [id, name, suppressed, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        place_id: { type: string, nullable: true }
        name: { type: string }
        category: { type: string, nullable: true }
        address: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        zip: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        website: { type: string, nullable: true }
        domain: { type: string, nullable: true }
        rating: { type: number, format: double, nullable: true }
        reviews: { type: integer, nullable: true }
        lat: { type: number, format: double, nullable: true }
        lng: { type: number, format: double, nullable: true }
        suppressed: { type: boolean }
        notes: { type: string, nullable: true }
        first_job_id: { type: string, format: uuid, nullable: true }
        first_job_name: { type: string, nullable: true, description: Name of the job that first found this business, null once that job is deleted }
        primary_email: { type: string, nullable: true }
        primary_email_source: { $ref: '#/components/schemas/EmailSource' }
        primary_email_verified_status:
          allOf:
            - $ref: '#/components/schemas/VerificationTag'
          nullable: true
          description: Tag of the primary address, null until it has been verified
        primary_email_verification_score: { type: integer, nullable: true }
        emails_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BusinessDetail:
      allOf:
        - $ref: '#/components/schemas/Business'
        - type: object
          properties:
            emails:
              type: array
              items: { $ref: '#/components/schemas/BusinessEmail' }
            raw:
              type: object
              additionalProperties: true
              nullable: true
    BusinessList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Business' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    BusinessUpdate:
      type: object
      properties:
        suppressed: { type: boolean }
        notes: { type: string, maxLength: 2000, nullable: true }
      additionalProperties: false
    BusinessBulkRequest:
      type: object
      required: [ids, action]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 5000
          items: { type: string, format: uuid }
        action:
          type: string
          enum: [suppress, unsuppress]
      additionalProperties: false
    BulkResult:
      type: object
      required: [updated]
      properties:
        updated: { type: integer }
    BusinessExportRequest:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        category: { type: string, maxLength: 200 }
        state: { type: string, maxLength: 100 }
        city: { type: string, maxLength: 200 }
        has_email: { type: boolean }
        suppressed: { type: boolean }
        q: { type: string, maxLength: 200 }
        email_source: { $ref: '#/components/schemas/EmailSource' }
        verification_tag:
          type: array
          items: { $ref: '#/components/schemas/VerificationTag' }
        ids:
          type: array
          maxItems: 5000
          items: { type: string, format: uuid }
          description: When present, exports exactly these rows and ignores the filters
      additionalProperties: false
...
    VerificationTag:
      type: string
      description: |
        The colour band a final score falls into. The UI must never convey a tag by
        colour alone; `tag_label` carries the text that goes with it.
      enum: [green, light_green, yellow, orange, red]
...
    ContactStage:
      type: string
      description: |
        A contact's position in the cold → subscriber funnel. The first eleven are the
        funnel proper and a contact only ever moves forward through them; the last six
        are terminal, outrank every funnel stage, and only `unsubscribed` can replace
        another terminal.
      enum: [cold, queued_for_instantly, contacted, engaged, replied, interested,
             permission_requested, permission_captured, newsletter_eligible,
             mailchimp_pending, mailchimp_subscribed,
             not_interested, wrong_person, do_not_contact, invalid_email, bounced, unsubscribed]
...
    ComponentType:
      type: string
      description: |
        The slots an email is assembled from, in assembly order. `subject` is its own
        line; the rest are joined with blank lines.
      enum: [subject, hook, problem, value_prop, proof, cta, closing, ps]
...
    Metrics:
      type: object
      description: |
        One bucket of send-level counts and the rates derived from them. Every rate is
        over `sends`, except `unsubscribe_rate`, which is over `unique_contacts`
        because a person unsubscribes once however many emails reached them.
      required: [sends, unique_contacts, opened, clicked, replied, positive_replies,
                 bounced, unsubscribed, open_rate, click_rate, reply_rate,
                 positive_reply_rate, bounce_rate, unsubscribe_rate]
      properties:
        sends: { type: integer, format: int64 }
        unique_contacts: { type: integer, format: int64 }
        opened: { type: integer, format: int64 }
        clicked: { type: integer, format: int64 }
        replied: { type: integer, format: int64 }
        positive_replies: { type: integer, format: int64 }
        bounced: { type: integer, format: int64 }
        unsubscribed: { type: integer, format: int64 }
        open_rate: { type: number, format: double }
        click_rate: { type: number, format: double }
        reply_rate: { type: number, format: double }
        positive_reply_rate: { type: number, format: double }
        bounce_rate: { type: number, format: double }
        unsubscribe_rate: { type: number, format: double }
    CountsByKey:
      type: object
      description: Counts keyed by a status, stage or other discriminator.
      additionalProperties: { type: integer, format: int64 }
...
    FunnelStep:
      type: object
      description: |
        One bar of the cold → subscriber funnel. `current` is who sits in the stage
        now, `ever_reached` who has passed through it; read the second as the
        conversion, because a contact only moves forward.
      required: [stage, label, current, ever_reached]
      properties:
        stage: { $ref: '#/components/schemas/ContactStage' }
        label: { type: string, example: Newsletter eligible }
        current: { type: integer, format: int64 }
        ever_reached: { type: integer, format: int64 }
    CampaignOverview:
      type: object
      required: [campaigns, contacts_by_stage, local, interested, newsletter_eligible,
                 subscribers, conversion_rate, funnel]
      properties:
        campaigns: { $ref: '#/components/schemas/CountsByKey' }
        contacts_by_stage: { $ref: '#/components/schemas/CountsByKey' }
        local: { $ref: '#/components/schemas/Metrics' }
        interested: { type: integer, format: int64 }
        newsletter_eligible: { type: integer, format: int64 }
        subscribers: { type: integer, format: int64 }
        conversion_rate: { type: number, format: double }
        funnel:
          type: array
          items: { $ref: '#/components/schemas/FunnelStep' }
    Mismatch:
      type: object
      description: A metric where our count and Instantly's disagree beyond tolerance.
      required: [metric, local, instantly]
      properties:
        metric: { type: string, example: replied }
        local: { type: integer, format: int64 }
        instantly: { type: integer, format: int64 }
    DailyPoint:
      type: object
      required: [day, sends, opened, clicked, replied, bounced]
      properties:
        day: { type: string, format: date }
        sends: { type: integer, format: int64 }
        opened: { type: integer, format: int64 }
        clicked: { type: integer, format: int64 }
        replied: { type: integer, format: int64 }
        bounced: { type: integer, format: int64 }
    CampaignAnalytics:
      type: object
      description: |
        Our numbers and Instantly's side by side. They are never merged: the two count
        differently, and the gap is itself a signal, which `mismatch` reports rather
        than averages away.
      required: [local, mismatch, daily, funnel, interested, eligible, subscribers,
                 conversion_rate, lead_counts]
      properties:
        local: { $ref: '#/components/schemas/Metrics' }
        instantly:
          type: object
          additionalProperties: true
          nullable: true
          description: Instantly's own figures, unchanged; null when they could not be fetched
        mismatch:
          type: array
          items: { $ref: '#/components/schemas/Mismatch' }
        daily:
          type: array
          items: { $ref: '#/components/schemas/DailyPoint' }
        funnel:
          type: array
          items: { $ref: '#/components/schemas/FunnelStep' }
        interested: { type: integer, format: int64 }
        eligible: { type: integer, format: int64 }
        subscribers: { type: integer, format: int64 }
        conversion_rate: { type: number, format: double }
        lead_counts: { $ref: '#/components/schemas/CountsByKey' }
    VariantAnalytics:
      allOf:
        - $ref: '#/components/schemas/Metrics'
        - type: object
          required: [variant_id, variant_name, step, assigned]
          properties:
            variant_id: { type: string, format: uuid }
            variant_name: { type: string }
            step: { type: integer, minimum: 1, maximum: 5 }
            assigned: { type: integer, format: int64 }
    ComponentAnalytics:
      allOf:
        - $ref: '#/components/schemas/Metrics'
        - type: object
          required: [component_id, type, name]
          properties:
            component_id: { type: string, format: uuid }
            type: { $ref: '#/components/schemas/ComponentType' }
            name: { type: string }
    AccountAnalytics:
      type: object
      required: [email, local, instantly_sent, instantly_bounced, instantly_replies,
                 instantly_bounce_rate]
      properties:
        sending_account_id: { type: string, format: uuid, nullable: true }
        email: { type: string }
        local: { $ref: '#/components/schemas/Metrics' }
        instantly_sent: { type: integer, format: int64 }
        instantly_bounced: { type: integer, format: int64 }
        instantly_replies: { type: integer, format: int64 }
        instantly_bounce_rate: { type: number, format: double }